% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{calc_variance}
\alias{calc_variance}
\title{Calculate the fraction of the variance of PCA restlts}
\usage{
calc_variance(pca)
}
\arguments{
\item{pca}{object returned by \code{\link[stats]{prcomp}}}
}
\value{
Vector of of the portion of variance expalained by the PCs.
}
\description{
Calculate the fraction of the variance of PCA restlts
}
\seealso{
\code{\link{plot_varfrac}}
}
