% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{check_columns}
\alias{check_columns}
\title{Check a dataframe or tibble for required columns, if missing columns throw an error}
\usage{
check_columns(input, req_cols)
}
\arguments{
\item{input}{the data frame or the tibble with the column names to check.}

\item{req_cols}{a vector of the required column name strings}
}
\description{
Check a dataframe or tibble for required columns, if missing columns throw an error
}
