% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gates.R
\name{chkgates}
\alias{chkgates}
\title{Find the runids that passed the gates in all experiments}
\usage{
chkgates(hdata, vars)
}
\arguments{
\item{hdata}{Data frame of hector ensemble runs, including all of the
experiments to use in the filtering}

\item{vars}{Names of variables to include in the filtering}
}
\value{
Vector of runids that passed all gates in all experiments.
}
\description{
Find the runids that passed the gates in all experiments
}
\examples{
ids <- chkgates(rbind(hector_conc_ensemble$rcp26, hector_conc_ensemble$rcp85),
               'tas')
length(ids)    # == 183
}
