% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cmip_individual}
\alias{cmip_individual}
\title{CMIP5 individual model comparison data}
\format{Data frame with 8 columns
\describe{
\item{year}{Year being described (1851--2300)}
\item{model}{Name of the model}
\item{ensemble}{Identifier for the ensemble member}
\item{variable}{Variable being reported ("tas" or "co2").  Temperatures are
anomalies relative to the first decade reported by the model.}
\item{experiment}{Experiment being reported. One of "historical", "rcp26", "rcp45",
"rcp60", "rcp85", "esmHistorical", "esmrcp85"}
\item{value}{Value of the variable being reported}
\item{esmbaseline}{Mean temperature (in Kelvin) for this model's esmHistorical run during
its first decade.  Temperatures for esmHistorical and esmrcp85 experiments
are reported as anomalies relative to this temperature.  This column will be
\code{NA} for mdoels that did not run the esm experiments.}
\item{concbaseline}{Mean temperature (in Kelvin) for this model's historical run during
its first decade.  Temperatures for concentration-driven runs are reported as
anomalies relative to this temperatrure.}
}}
\usage{
cmip_individual
}
\description{
This dataset contains outputs for the individual models in the ensemble.
}
\seealso{
Other comparison data: \code{\link{esm_comparison}}
}
\concept{comparison data}
\keyword{datasets}
