% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{compute_pc}
\alias{compute_pc}
\title{Compute principal components from an ensemble of Hector runs}
\usage{
compute_pc(scenlist, experiments, variables, years, histyears,
  retx = FALSE)
}
\arguments{
\item{scenlist}{List of data frames of hector runs, one for each experiment.
The names of the list elements should correspond to the experiments contained
in each.}

\item{experiments}{Vector of strings denoting the experiments to include.
Elements of scenlist not matching one of these strings will be dropped.}

\item{variables}{Vector of Hector output variables to include in the analysis.}

\item{years}{Vector of years to include in experiments using \emph{future}
runs.  Generally these will be the various RCP pathways.}

\item{histyears}{Vector of years to include in experiments using \emph{past}
runs.  Any experiment with ``historical'' or ``Historical'' in the name is
considered to be a past run.}

\item{retx}{Flag indicating whether to retain the projections of the ensemble
runs onto the principal components.}
}
\description{
Given a collection of Hector runs, filter to the desired experiments,
variables, years, and future years and compute the principal components for
that combination.
}
\details{
This function will work with the package datasets
\code{\link{hector_conc_ensemble}} and \code{\link{hector_emiss_ensemble}}
as the \code{scenlist} argument.

On the ordering of variables when constructing principal components: When
flattening a set of runs into a vector, the grouping order shall be
experiment, variable, year.  That is, year will vary the most rapidly,
experiment the least.  This ordering is assumed in all of the other functions
that deal with principal components.
}
