% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc-helpers.R
\name{concat_runs}
\alias{concat_runs}
\title{Concatenate a set of base runs and continuations into a single run structure.}
\usage{
concat_runs(bruns, cruns, output_dir = NULL,
  name_stem = "hectorcal_concat", codasize = 2500)
}
\arguments{
\item{bruns}{List of Monte Carlo results from the base runs.}

\item{cruns}{List of Monte Carlo results from the continuation runs.}

\item{output_dir}{Optional output directory to write concatenated objects into.}

\item{name_stem}{String giving the stem of the filenames for saving concatenated objects.}

\item{codasize}{Number of samples to keep in the \code{\link[coda]{mc.list}} objects in the output.}
}
\value{
Structure as in \code{\link{proc_mc_rslts}}, for the concatenated runs
}
\description{
The inputs to this function are two sets of outputs from \code{\link{proc_mc_rslts}}.  The
function will extract the metrosamp objects from the two outputs, concatenate them using
the \code{\link[metrosamp]{concat}} function, and produce the \code{runstats} and \code{codaobj}
structures erturned by \code{proc_mc_rslts}.  Optionally, write the concatenated results to
files so that they can be read in directly in the future.
}
