% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{difference_between}
\alias{difference_between}
\title{Determine the elements that are different between two vectors.}
\usage{
difference_between(x, y)
}
\arguments{
\item{x}{a vector}

\item{y}{a second vector}
}
\value{
a vector containing the elements that are unique to the two input vectors
}
\description{
Determine the elements that are different between two vectors.
}
