% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{encode_runid}
\alias{encode_runid}
\alias{decode_runid}
\alias{load_mc_output}
\title{Encode and decode runids, and find matching runs.}
\usage{
encode_runid(serialnumber, pcsflag, hfflag, meanflag, debugflag)

decode_runid(runid)

load_mc_output(pcsflag, hfflag, meanflag, debugflag, dir = ".",
  filestem = "hectorcal", niter = NA)
}
\arguments{
\item{serialnumber}{Vector of serial numbers.  Serial numbers must be between
0 and 15, inclusive}

\item{pcsflag}{Vector of flags indicating whether or not to use principal
components.}

\item{hfflag}{Vector of flags indicating whether or not to use heat flux}

\item{meanflag}{Vector of flags indicating whether or not to do mean
calibration (as opposed to envelope calibration)}

\item{debugflag}{Vector of flags indicating whether or not to include
debugging information.}

\item{runid}{Vector of integer runid values}

\item{dir}{Directory to load files from}

\item{filestem}{Stem of the data file names}

\item{niter}{Number of iterations. If NA, match all iteration counts}
}
\value{
Vector of runid values

List containing a vector of serial numbers and logical vectors for
each of the four flag types.
}
\description{
Functions to convert runids to serial numbers and flags, and vice versa, and
to load data matching specified runids or flags.
}
\details{
Definition of runid:
\describe{
 \item{Bits 0-3:}{Serial number (used to give each run of the same
configuration a unique RNG seed}
 \item{Bit 4:}{Use pcs flag.  If false, we use the outputs directly.}
 \item{Bit 5:}{Use heat flux flag.}
 \item{Bit 6:}{Mean calibrtion flag}
 \item{Bit 7:}{Debug flag}
}
}
\section{Functions}{
\itemize{
\item \code{decode_runid}: Decode a vector of runids into flags and serial numbers

\item \code{load_mc_output}: Load available runs for a given configuration.
}}

