% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{esm_comparison}
\alias{esm_comparison}
\title{CMIP5 ensemble comparison data}
\format{Data frame with 9 columns
\describe{
\item{year}{Year being described (1851--2300)}
\item{variable}{Variable being reported ("tas" or "co2").  Temperatures are
given as anomalies relative to the first decade reported by the model.}
\item{experiment}{Experiment being reported.  One of "historical", "rcp26", "rcp45",
"rcp60", "rcp85", "esmHistorical", "esmrcp85"}
\item{mina}{Minimum value reported in the ensemble}
\item{maxb}{Maximum value reported in the ensemble}
\item{a10}{First decile reported in the ensemble}
\item{b90}{Ninth decile reported in the ensemble}
\item{cmean}{Mean value reported in the ensemble}
\item{cmedian}{Median value reported in the ensemble}
}}
\usage{
esm_comparison
}
\description{
This dataset contains summary statistics for the CMIP5 ensemble results by year,
variable, and experiment.
}
\section{Notes}{


No effort was made to account for the fact that some models have more
ensemble members in an experiment than others, nor for the fact that some
models have several (presumably closely related) variants in the ensemble
(e.g., CESM has CESM1-BGC, CESM1-CAM5, CESM1-FASTCHEM, and CESM1-WACCM
represented).  For extreame statistics like min, max, and probably the decile
points, this doesn't make a big difference, but central statistics like mean and
mdian, the difference in representation could give some models undue
influence on how the statistics turn out.
}

\seealso{
Other comparison data: \code{\link{cmip_individual}}
}
\concept{comparison data}
\keyword{datasets}
