% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{esm_pcdata}
\alias{esm_pcdata}
\alias{cmip_conc_pcproj}
\alias{cmip_emiss_pcproj}
\title{Principal component projection coefficients for ESM outputs}
\format{A data frame with 3 columns
\describe{
\item{model}{Name of the model.}
\item{PC}{Index of the principal component.}
\item{value}{Projection coefficient.}
}}
\usage{
cmip_conc_pcproj

cmip_emiss_pcproj
}
\description{
These tables give the projection coefficients onto the hector principal
components for each of the CMIP5 models that provided sufficient data to make
the projection.  The concentration-driven runs are projected onto
\code{\link{pc_conc}}, and the emissions-driven runs are projected onto
\code{\link{pc_emiss}}.
}
\seealso{
Other comparison data: \code{\link{cmip_individual}},
  \code{\link{esm_comparison}}, \code{\link{pccd}}
}
\concept{comparison data}
\keyword{datasets}
