% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{esm_pcplot}
\alias{esm_pcplot}
\title{Plot the first two PC coordinates for a table of ESM projections}
\usage{
esm_pcplot(pctable)
}
\arguments{
\item{pctable}{Data frame containing the ESM projections by model}
}
\description{
Plot the first two PC coordinates for a table of ESM projections
}
