% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{find_common_years}
\alias{find_common_years}
\title{For a given experiment, find all of the common years for each variable.}
\usage{
find_common_years(df, expt)
}
\arguments{
\item{df}{Data frame of model runs.  Should have same format as \code{\link{cmip_individual}}.}

\item{expt}{Name of the experiment (i.e., scenario) to examine.}
}
\value{
Named list of common years by variable
(i.e., \code{list(tas=tas_years, co2=co2_years, ...)})
}
\description{
"Common years" are the years that are present in all models and ensemble members
in the dataset.
}
