% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleESMcalibration.R
\name{generate_inital_guess}
\alias{generate_inital_guess}
\title{Use hectorcal's ensemble of Hector of hector results to guess an inital combination of parameter values
to use as tge best guess in the single ESM calibration package.}
\usage{
generate_inital_guess(comparison_data, param_names)
}
\arguments{
\item{comparison_data}{A dataset set containing the comparison data that is going to be used by the
calibration functions. Should contain the following columns: year, variable, value.}

\item{param_names}{A vector of the Hector parameter names to find the inital paramter value guesses for.}
}
\value{
A vector of the inital guesses for the Hectpr parameters being calibrated by
 \code{singleESM_calibration} and \code{singleESM_calibration_diag}.
}
\description{
Some of the paramter values that were un realistic, extreme values (i.e., alpha values less than 0, S greater than 30 ect.)
 are excluded from the analysis.
}
