% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{get_gates}
\alias{get_gates}
\title{Get a data frame of gates for an experiment}
\usage{
get_gates(expt, var, minyear = 1861, maxyear = 2100)
}
\arguments{
\item{expt}{Experiment to retrieve ('historical', 'rcp26', 'rcp45', etc.)}

\item{var}{Variable to retrieve ('tas' or 'co2')}

\item{minyear}{Earliest year to retrieve.}

\item{maxyear}{Latest year to retrieve.}
}
\value{
Data frame with the matching rows from the
\code{\link{esm_comparison}} dataset.
}
\description{
Find the minimum and maximum attested values by year for a climate variable
in the CMIP5 output.  Values are returned at the beginning, end, and median
of the requested period.
}
\details{
For historical experiments, the final year can be no later than 2005.  For
future experiments, the starting year can be no earlier than 2006.
The \code{minyear} and \code{maxyear} parameters will be adjusted
automatically to these limits.
}
