% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hector_emiss_ensemble}
\alias{hector_emiss_ensemble}
\title{Ensembles of Hector emission driven runs}
\format{List of 6 data frames: esmHistorical, esmrcp26, esmrcp45, esmrcp60,
esmrcp85, and params.  The four output data frames have 6 columns:
\describe{
\item{runid}{Unique identifier for each run.}
\item{variable}{Variable being reported "tas", temperatures are anomalies.}
\item{scenario}{Hector scenario name.}
\item{experiment}{Experiment being reported, should match the cmip experiments. One of  "rcp26", "rcp45", "rcp60", "rcp85", "esmHistorical", "esmrcp85".}
\item{year}{Year being described (1850 -- 2100).}
\item{value}{Value of the variable being reported.}
}
The params data frame has 7 columns
\describe{
\item{runid}{Unique identifier for each run.}
\item{S}{Equilibrium climate sensitivity.}
\item{alpha}{Aerosol scaling factor.}
\item{volscl}{Volcanic forcing scaling factor.}
\item{diff}{Diffusivity parameter.}
\item{C0}{Preindustrial CO2 concentration.}
\item{q10_rh}{Heterotrophic respiration Q10 parameter.}
\item{beta}{CO2 fertilization parameter}
}}
\usage{
hector_emiss_ensemble
}
\description{
This dataset contains outputs for the 1000 Hector runs driven with each of
the four concentration pathways, along with the parameters used to drive the
runs.
}
\seealso{
Other PCA hector ensemble: \code{\link{hector_conc_ensemble}}
}
\concept{PCA hector ensemble}
\keyword{datasets}
