% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{hector_output_stats}
\alias{hector_output_stats}
\title{Produce summary statistics for hector outputs}
\usage{
hector_output_stats(params, hcores, pnames, times = c(2000, 2050, 2100),
  nsamp = NULL, quantiles = c(0, 0.025, 0.25, 0.5, 0.75, 0.975, 1))
}
\arguments{
\item{params}{Matrix of parameter samples, parameters in columns, samples in
rows.}

\item{hcores}{A hector core, or a list of hector cores, configured for the
appropriate emissons.  If a list of cores is supplied, the samples will be
run in parallel}

\item{pnames}{Vector of hector parameter names corresponding to the columns
in the params array.}

\item{times}{Vector of times to sample.}

\item{nsamp}{Number of samples to take.  More samples gives more accuracy,
but takes longer.  Default is greater of 100 or 10% of the number of input
samples.}

\item{quantiles}{Quantiles to report}
}
\value{
matrix of quantile values for each time and variable
importFrom foreach foreach %do% %dopar%
}
\description{
For a matrix of samples of parameter values, produce summary statistics for
the hector outputs generated by those parameters
}
