% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{likelihoodInternal}
\alias{likelihoodInternal}
\alias{make_logprior}
\alias{make_loglikelihood}
\alias{make_logpost}
\title{Internal functions for creating log-posterior functions}
\usage{
make_logprior(pprior, use_lnorm_ecs)

make_loglikelihood(inifiles, verbose, cal_mean, comp_data, hicol, lowcol,
  pcs, hflux_year, hflux_expt_regex)

make_logpost(lprior, llik, verbose)
}
\arguments{
\item{pprior}{Parameters for prior distributions.  This \emph{must} be a
complete set (unlike the input to \code{build_mcmc_post})}

\item{use_lnorm_ecs}{Flag indicating whether to use the lognormal prior for
climate sensitivity}

\item{inifiles}{Named vector of names of ini files for the various
experiments. You do not need to provide ini files for historical
experiments unless they are the only experiments present.}

\item{verbose}{Flag indicating whether diagnostic messages should be}

\item{cal_mean}{Flag indicating whether we are calibrating to the mean. If
\code{FALSE}, calibrate to the envelope.}

\item{comp_data}{Observed comparison data, in summary format (see, for
example, \code{\link{esm_comparison}}.}

\item{hicol}{Name of the column with the high edge of the window.}

\item{lowcol}{Name of the column with the low edge of the window.}

\item{pcs}{Principal components structure.  If \code{NULL} outputs will be
compared directly.}

\item{hflux_year}{Year at which to evaluate ocean heat flux for comparison.}

\item{hflux_expt_regex}{Regular expression identifying the experiment to
get heat flux from.}

\item{lprior}{Function that computes the log-prior}

\item{llik}{Function that computes the log-posterior}

\item{verbose}{Flag indicating whether diagnostic output should be produced}
}
\description{
These functions are called by \code{\link{build_mcmc_post}} to create the
log-posterior functions
}
\section{Functions}{
\itemize{
\item \code{make_logprior}: Make a log prior function

\item \code{make_loglikelihood}: Make a log likelihood function

\item \code{make_logpost}: Combine prior and likelihood into a single function.
}}

\keyword{internal}
