% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleESMcalibration.R
\name{make_minimize_function}
\alias{make_minimize_function}
\title{Make the function to minimize}
\usage{
make_minimize_function(hector_cores, esm_data, normalize, param)
}
\arguments{
\item{hector_cores}{A list of Hector cores that will be used.}

\item{esm_data}{A data frame of ESM data for a single model that contains the following columns, year, model, variable, experiment.}

\item{normalize}{A list of center and the scale values to use to noramlize the Hector and ESM output data.}

\item{param}{A vector of parameter values.}
}
\value{
A function that will calculate the mean squared error between a Hector run and ESM data.
}
\description{
Create the function that will calculate the mean squared error between Hector and ESM CMIP5 output data. THis
function will be minimized using optim.
}
