% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleESMcalibration.R
\name{make_param_penalty_function}
\alias{make_param_penalty_function}
\title{Make the parameter pentalty function}
\usage{
make_param_penalty_function(func_list)
}
\arguments{
\item{func_list}{a list of named functions contain the penalty function for the parameters to penalize.}
}
\value{
A function that will return a tibble containing the penalty for each penalized Hector parameter
in a data frame.
}
\description{
Create the parameter pentalty function that is used to penalize the optmized value in the single ESM
calibration for specific parameter values. The function returned by \code{make_param_penalty_function}
is used within the \code{make_minimize_function}.
}
