% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleESMcalibration.R
\name{make_param_penatlity_function}
\alias{make_param_penatlity_function}
\title{Make a paramter pentalty function}
\usage{
make_param_penatlity_function(penalize, lower, upper, sig)
}
\arguments{
\item{penalize}{A vector of the Hector parameter names to penalize.}

\item{lower}{A vector of the lower boundries for the parmeters listed in penalize}

\item{upper}{A vector of the upper boundries for the parmeters listed in penalize}

\item{sig}{A vector of the parameter controlling the steepness of the rolloff at the boundaries}
}
\value{
A function that will return a tibble containing the -log mesa penalty for the Hector parameter by epxeriment and variable or parameter in this case.
}
\description{
Create a function that will penalize mean squared error between Hector and ESM CMIP5 output data based on the
certain parameter values. The function returned by \code{make_param_penatlity_function} is used within the \code{make_minimize_function}.
This function judges Hector paramter fits based on the - log \code{mesa} function.
}
