% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleESMcalibration.R
\name{make_parameterize_cores}
\alias{make_parameterize_cores}
\title{Make a function that will parameterize a list of Hector cores with new Hector parameters.}
\usage{
make_parameterize_cores(param)
}
\arguments{
\item{param}{a vector of named Hector parameters to used to set up the cores.}
}
\value{
A function that can be used to reset a sinlge Hector core with new Hector parameters.
}
\description{
The function
}
