% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc-helpers.R
\name{makeplots}
\alias{makeplots}
\title{Make diagnostic plots for Monte Carlo Runs}
\usage{
makeplots(plotruns, tofile = TRUE, filestem = "hectorcal-run-")
}
\arguments{
\item{plotruns}{List of metrosamp objects for runs in a single configuration.}

\item{tofile}{Flag indicating whether the result should be written to a file}

\item{filestem}{String giving the stem of the filename to be written.}
}
\description{
Making these plots is very slow for large runs.  Mostly they are
useful for diagnosing pilot runs
}
\keyword{internal}
