% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesa.R
\name{mesa}
\alias{mesa}
\title{Calculate the mesa function used in defining our likelihoods.}
\usage{
mesa(x, a, b, sig)
}
\arguments{
\item{x}{Values at which to evaluate the mesa function}

\item{a}{Parameter controlling the lower boundary of the function.}

\item{b}{Parameter controlling the upper boundary of the function.}

\item{sig}{Parameter controlling the steepness of the rolloff at the
boundaries.}
}
\description{
The mesa function is designed to be flat over some range and then to fall off
more or less sharply as it approaches the edges of that range.  Outside the
range, it has Gaussian-like tails.
}
\details{
The parameters \code{a} and \code{b} control the width of the flat range.
One notable feature of the function is that the probability density at x=a or
x=b is always half what it is in the flat region of the function.

The \code{sig} parameter controls how rapidly the probability density falls
off as you approach the boundary.  Larger values of \code{sig} extend the flat
region of the function closer to the boundary, and causes the rolloff to be
sharper, and the tails outside the boundary weaker.  Smaller values cause the
rolloff to start farther from the boundary, make the rolloff more gradual,
and cause the tails to extend farther outside the boundary.

The function is vectorized, so it is permissible to pass a vector of x
values.  The other parameters will be recycled as vectors in the usual way.
}
