% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc-helpers.R
\name{pairplot}
\alias{pairplot}
\title{Plot pairs plots for the samples in a Monte Carlo configuration.}
\usage{
pairplot(mcrslt)
}
\arguments{
\item{mcrslt}{List of metrosamp objects for runs in a configuration}
}
\value{
The correlation matrix for the samples
}
\description{
The pairs plots can be used to get some insight into the joint distribution
of the parameters.
}
