% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleESMcalibration.R
\name{parameterize_core}
\alias{parameterize_core}
\title{Parameterize a Hector core.}
\usage{
parameterize_core(params, core)
}
\arguments{
\item{params}{A named vector of Hector parameters.}

\item{core}{A Hector core.}
}
\value{
A Hector set up to run with a new set of paramters.
}
\description{
Parameterize a Hector core.
}
