% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pc_conc_hflux}
\alias{pc_conc_hflux}
\title{Hector concentration driven principal compoents (PCs) with heat flux}
\format{A list with 5 elements
\describe{
\item{sdev}{A vector of the standard deviations of the principal components.}
\item{rotation}{A matrix of the variable loadings.}
\item{center}{The centering used in the PCA.}
\item{scale}{The scale used in the PCA.}
\item{meta_data}{a list of meta data information about the data that went into the PCA}
}
The metadata element is itself a list with four elements:
\describe{
\item{year}{Year used in the PCA (1850 -- 2100).}
\item{variable}{The variables that describe the values used in the PCA ("tas"
and "heatflux").}
\item{experiment}{The CMIP5 experiment names describing the data that was used in the PCA.}
\item{scenario}{The Hector scenario name describing the data that was used in the PCA.}
}}
\usage{
pc_conc_hflux
}
\description{
This oject contains the sd, rotation, center, and scale for the PCs calculated with
\code{prcomp} on an enemsble of Hector concentration driven RCP 2.6, RCP 4.5, RCP 6.0, and RCP 8.5 runs
along with metadata about the Hector data that went into the runs.
}
\seealso{
Other PC hector: \code{\link{pc_conc}},
  \code{\link{pc_emiss_hflux}}, \code{\link{pc_emiss}}
}
\concept{PC hector}
\keyword{datasets}
