% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PC_hector_concen-rcp26_concen-rcp45_concen-rcp60_concen-rcp85}
\alias{PC_hector_concen-rcp26_concen-rcp45_concen-rcp60_concen-rcp85}
\title{Hector concentration driven principal compoents (PCs)}
\format{A list with 5 elements
\describe{
\item{sdev}{A vector of the standard deviations of the principal componets.}
\item{rotation}{A matrix of the variable loadings.}
\item{center}{The centering used in the PCA.}
\item{scale}{The scale used in the PCA.}
\item{meta_data}{a list of meta data information about the data that went into the PCA}
\item{year}{Year used in the PCA (1850 -- 2100).}
\item{variable}{The variables that describe the values used in the PCA ("tas").}
\item{experiment}{The CMIP5 experiment names describing the data that was used in the PCA.}
\item{scenario}{The Hector scenario name describing the data that was used in the PCA.}
}}
\usage{
PC_hector_concen-rcp26_concen-rcp45_concen-rcp60_concen-rcp85
}
\description{
This oject contains the sd, rotation, center, and scale for the PCs calculated with
\code{prcomp} on an enemsble of Hector concentration driven RCP 2.6, RCP 4.5, RCP 6.0, and RCP 8.5 runs
along with some meta data information about the Hector data that went into the runs
}
\seealso{
Other PC hector: \code{\link{PC_hector_emissCC-esmrcp26_emissCC-esmrcp45_emissCC-esmrcp60_emissCC-esmrcp85}},
  \code{\link{PC_hector_emissConstantC-esmrcp26_emissConstantC-esmrcp45_emissConstantC-esmrcp60_emissConstantC-esmrcp85}}
}
\concept{PC hector}
\keyword{datasets}
