% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{pc_rotplot}
\alias{pc_rotplot}
\title{Plot the rotation of the principal components into the output space.}
\usage{
pc_rotplot(pclimits, pcstruct, yrx = 2006, yry = 2100, var = "tas",
  expt = "rcp85", pccords = NULL, HML = c("max", "mean", "min"))
}
\arguments{
\item{pclimits}{Data frame giving the high, medium, and low values for each
PC.}

\item{pcstruct}{Principal components structure.}

\item{yrx}{Year to plot as the x-value in the output space}

\item{yry}{Year to plot as tye y-value in the output space}

\item{var}{Variable to plot in the output space}

\item{expt}{Experiment to use for the output variable assignment.}

\item{pccords}{Assignment of coordinates to principal components (see details)}

\item{HML}{Names of the columns in pclimits containing the High, Medium, and
Low values for the gates.  (Also, my grandmother's initials.  Miss you.)}
}
\description{
Show what a gate defined in the principal components space looks like in the
output space.
}
\details{
Six principal components can be used in the plot.  Two of them will be used
to draw the boxes in the output space.  The other four will be assigned to
the color, linetype, facet row, and facet column aesthetics.  These
assignments are referred to as x, y, col, lty, facetrow, and facetcol,
respectively.  By default they are assigned to PCs 1-6 in that order.  This
assignment can be changed by passing a named vector to the pccords
parameter.  The names give the roles to rebind, and the (integer) values
give the PC to bind to them.  You only need to pass the ones you want to change
from their defaults, so for example, passing \code{c(x=3, lty=1)} would swap
the roles of PC1 and PC3, while leaving the other roles unchanged.

Note that if any of the PCs used in pccords are not in the pclimits table, or
if any of them occur more than once, you will get an error.
}
