% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PCA_ESMmean-concen}
\alias{PCA_ESMmean-concen}
\title{Concentration driven CMIP5 ESM mean comparison data}
\format{Data frame with 4 columns
\describe{
\item{year}{Year being described (1851 -- 2100).}
\item{variable}{Variable being reported "tas", temperatures are anomalies.}
\item{experiment}{Experiment being reported, "esmrcp85"}
\item{value}{The multimodel mean.}
}}
\usage{
PCA_ESMmean-concen
}
\description{
This dataset contains the CMIP5 ensmeble mean output for the concentration driven historical and
rcp85 experiments. The multi-model mean is the average of each model intital conditions ensemble
average. This multi-model mean does not account for how models may be related to one aother. The
historical experiment name has been replaced with the rcp experiment name.
}
\seealso{
Other PCA hector ensemble: \code{\link{PCA_ESMmean-emiss}},
  \code{\link{PCA_hector_ensemble-concen-rcp26}},
  \code{\link{PCA_hector_ensemble-concen-rcp45}},
  \code{\link{PCA_hector_ensemble-concen-rcp60}},
  \code{\link{PCA_hector_ensemble-concen-rcp85}},
  \code{\link{PCA_hector_ensemble-emissConstantC-esmrcp26}},
  \code{\link{PCA_hector_ensemble-emissConstantC-esmrcp45}},
  \code{\link{PCA_hector_ensemble-emissConstantC-esmrcp60}},
  \code{\link{PCA_hector_ensemble-emissConstantC-esmrcp85}}
}
\concept{PCA hector ensemble}
\keyword{datasets}
