% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PCA_hector_ensemble-emissConstantC-esmrcp85}
\alias{PCA_hector_ensemble-emissConstantC-esmrcp85}
\title{Ensemble of Hector rcp85 emission driven runs with constant carbon cycle parameters}
\format{Data frame with 9 columns
\describe{
\item{runid}{Contains integers to indicate that results are from the same Hector run}
\item{variable}{Variable being reported "tas", temperatures are anomalies.}
\item{scenario}{Hector scenario name.}
\item{experiment}{Experiment being reported, should match the cmip experiments. One of  "rcp26", "rcp45", "rcp60", "rcp85", "esmHistorical", "esmrcp85".}
\item{year}{Year being described (1850 -- 2100).}
\item{value}{Value of the variable being reported.}
\item{S}{Value of the Hector run climate sensitivity.}
\item{alpha}{Value of the Hector run aerosol scaling factor.}
\item{diff}{Value of the Hector run ocean heat diffusivity.}
}}
\usage{
PCA_hector_ensemble-emissConstantC-esmrcp85
}
\description{
This dataset contains outputs for the 1000 Hector runs driven with rcp 8.5 emissions, this
data set only contains results from runs that also solved in for the Hector emissions driven
rcp 4.5, rcp 6.0, and rcp 8.5 runs. The carbon cycle parameters were held constant here.
}
\seealso{
Other PCA hector ensemble: \code{\link{PCA_hector_ensemble-concen-rcp26}},
  \code{\link{PCA_hector_ensemble-concen-rcp45}},
  \code{\link{PCA_hector_ensemble-concen-rcp60}},
  \code{\link{PCA_hector_ensemble-concen-rcp85}},
  \code{\link{PCA_hector_ensemble-emissConstantC-esmrcp26}},
  \code{\link{PCA_hector_ensemble-emissConstantC-esmrcp45}},
  \code{\link{PCA_hector_ensemble-emissConstantC-esmrcp60}}
}
\concept{PCA hector ensemble}
\keyword{datasets}
