% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfuns.R
\name{plot_frac}
\alias{plot_frac}
\title{Plot the fraction of variance captured by the PCs}
\usage{
plot_frac(frac, title = NULL)
}
\arguments{
\item{frac}{a vector of cumulative variance values}

\item{title}{optional string for plot title}
}
\value{
ggplot dot plot of the fraction of the variance
}
\description{
Plot the fraction of variance captured by the PCs
}
