% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_pc}
\alias{plot_pc}
\title{Make plots of principal components}
\usage{
plot_pc(pc, nc = NA, vartot = 0.995, rscl = FALSE,
  swap_layout = FALSE)
}
\arguments{
\item{pc}{Principal components structure}

\item{nc}{Number of components to plot.  Default is to plot up to
\code{vartot} total variance.}

\item{vartot}{Cumulative variance cutoff for determining how many components
to plot.  Ignored if \code{nc} is specified.}

\item{rscl}{If true, reverse the rotation and scale factors applied by the
PCA. This will have the effect of displaying the variables in their natural
units.}

\item{swap_layout}{Swap the layout of the facet grid from the default.
Ignored if \code{rscl} is true, since then we need to have variables in
columns to make the scales come out right.}
}
\value{
ggplot plot structure.
}
\description{
Make plots of principal components
}
