% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfuns.R
\name{plot_pcs}
\alias{plot_pcs}
\title{Plot N PCs from the object returend from prcomp}
\usage{
plot_pcs(pca, pc_indx, title = NULL)
}
\arguments{
\item{pca}{the PCA results, an object returned by \code{\link[stats]{prcomp}}}

\item{pc_indx}{vector of the PC indices to plot}

\item{title}{optional title for the plot}
}
\value{
ggplot line plot of the PCs
}
\description{
Plot N PCs from the object returend from prcomp
}
