% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{project_climate}
\alias{project_climate}
\title{Project a data frame of climate data on to a principal componet basis}
\usage{
project_climate(climate_data, principal_components)
}
\arguments{
\item{climate_data}{an object of climate data, must contain value, year, experiment, and variable columns.
It may not contain data for more years, variables, or experiments then in the principal_components argument.}

\item{principal_components}{a \code{hectorcal} object that contains results from \code{prcomp} and meta data information.}
}
\value{
a list of three objects, rotation the PC loadings, scale the values used to scale the input climate data, and center the values used to center the climate data.
}
\description{
Project a data frame of climate data on to a principal componet basis
}
