% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{run_hector_samples}
\alias{run_hector_samples}
\title{Run hector runs for sampled parameters}
\usage{
run_hector_samples(allsamps, nsamp, hcores, times, pnames, vars)
}
\arguments{
\item{allsamps}{Matrix of all parameter samples}

\item{nsamp}{Number of samples to draw.  Must not be greater than
\code{nrow(allsamps)}.}

\item{hcores}{List of hector cores initialized with the desired emissions.}

\item{times}{Times to pull from the hector output.}

\item{pnames}{Hector internal names for the parameters in the columns of
\code{allsamps}.}

\item{vars}{Variables to retrieve.}
}
\description{
This function is used as a helper function in some of our other plotting
function, but it is also useful in its own right.  For example, the data
frame returned by this function is just what you need to plot density plots
of the Hector outputs
}
