% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleESMcalibration.R
\name{setup_hector_cores}
\alias{setup_hector_cores}
\title{Make a list of Hector cores}
\usage{
setup_hector_cores(inifile, name)
}
\arguments{
\item{inifile}{A vector of ini file paths that will be used to set up the Hector cores}

\item{name}{A vector of the Hector core scenario names, it should reflect CMIP5 experiment names.}
}
\value{
A list of hector instances
}
\description{
The list of Hector cores will be passed into the \code{make_fn_minimize}
}
