% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleESMcalibration.R
\name{singleESM_calibration_diag}
\alias{singleESM_calibration_diag}
\title{Emulate a single ESM by calibrating Hector to ESM output data and return diagnostic materials.}
\usage{
singleESM_calibration_diag(inifiles, hector_names, esm_data, normalize,
  initial_param, cmip_range = NULL, param_penalty = NULL,
  maxit = 500, n_parallel = NULL, showMessages = FALSE)
}
\arguments{
\item{inifiles}{A vector of the Hector inifile cores to use, there should be a core for each esm CMIP comparison data.}

\item{hector_names}{A vector of the Hector core names, should reflect experiment names in the esm CMIP comparion data.}

\item{esm_data}{A data frame of ESM data for a single model that contains the following columns, year, model, variable, experiment.}

\item{normalize}{A list of center and the scale values to use to noramlize the Hector and ESM output data.}

\item{initial_param}{A named vector of inital paramters to be optimized over.}

\item{cmip_range}{Default set to NULL and the make_minim_function will only look at the error between esm_data and Hector data.
But if cmip_range is set to a dataframe containing columns (variable, year, sig, lower, and upper) then the minimize function will also
minimize the value returned by the -log of the mesa function.
 @param param_penalty Default is set to NULL but can be set to a function that will return a tibble of experiment / variable / value to penalize the
mean squared error for "unreal" paramter values.}

\item{maxit}{The max number of itterations for optim, default set to 500.}

\item{n_parallel}{The max number of cores to parallize the runs over,  unless sepcified will use the number of cores detected by \code{detectCores}.}

\item{showMessages}{Default set to FALSE, will supress Hector error messages.}
}
\value{
A list containing the following elements, copmarison_plot a plot comparing Hector and ESM output data, residual_plot a plot comparing the
normalized residuals, MSE a data frame of the mean squared error for each experiment and variable optim minimizes the sum of the MSE values, and
optim_rslt is the object returned by \code{optim}.
}
\description{
Calibrate Hector so that is emulates a single ESM. Calcaulte the mean squared error between the calibrated Hector and the
ESM comparison data, plot the Hector output vs the ESM comparison data, and a plot of the residuals.
}
