% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{spaghetti_plot}
\alias{spaghetti_plot}
\title{Plot spaghetti plots for hector samples}
\usage{
spaghetti_plot(mcrslt, nplot, hcores, pnames,
  vars = c(hector::GLOBAL_TEMP(), hector::ATMOSPHERIC_CO2()),
  times = 1850:2100, alpha = 0.3)
}
\arguments{
\item{mcrslt}{List of metrosamp structures from a Bayesian Monte Carlo run}

\item{nplot}{Number of traces to plot.}

\item{hcores}{List of hector cores initialized with the desired emissions.}

\item{pnames}{Hector names of the parameters in the sample matrix}

\item{vars}{Variables to plot}

\item{times}{Vector of times to include in the plot.  Default is 1850-2100.}

\item{alpha}{Alpha channel (transparency) for the sample traces.}
}
\description{
Spaghetti plots show the traces of hector outputs for a set of samples from a
Bayesian Monte Carlo calculation.
}
