% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{summarize_pcdecomp}
\alias{summarize_pcdecomp}
\title{Create comparison data for principal components decomposition}
\usage{
summarize_pcdecomp(pcdecomp, pcmax = NULL, pcselect = NULL)
}
\arguments{
\item{pcmax}{Maximum PC to use.  PCs higher than this will be dropped.}

\item{pcselect}{Individual PCs to select.  This is intersected with the list
implied by pcmax.}

\item{pcdata}{Input principal components decomposition (probably either
\code{\link{cmip_conc_pcproj}} or \code{\link{cmip_emiss_pcproj}}).}
}
\description{
We have stored the principal components decomposition for each CMIP5 model as
package data.  This function summarises that into a format similar to the
\code{\link{esm_comparison}} dataset.
}
