% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{update_pca_struct}
\alias{update_pca_struct}
\title{Update PCA structure metadata to the new format}
\usage{
update_pca_struct(pcastruct)
}
\arguments{
\item{pcastruct}{Old format PCA structure}
}
\value{
New format PCA structure
}
\description{
The new format for PCA structures has lists instead of vectors for the years and
histyears elements.  Use this function when you have a structure saved in the old
format and you need to convert it to the new.
}
