% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{vartranslate}
\alias{vartranslate}
\alias{hvar2esmvar}
\alias{esmvar2hvar}
\title{Translate between Hector and ESM variable names}
\usage{
hvar2esmvar(hvarnames)

esmvar2hvar(esmvarnames)
}
\arguments{
\item{hvarnames}{Vector of Hector variable names}

\item{esmvarnames}{Vector of ESM variable names}
}
\value{
Translated names. Unknown names will be replaced with \code{NA}.
}
\description{
Translate hector variable names ("Tgav", etc.) to ESM names and vice versa.
}
\section{Functions}{
\itemize{
\item \code{hvar2esmvar}: Translate hector variable names ('Tgav', etc.) to ESM names.

\item \code{esmvar2hvar}: Translate ESM variable names ('tas', etc.) to hector names.
}}

