% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{build_mcmc_post}
\alias{build_mcmc_post}
\title{Construct a log posterior probability function for Bayesian calibration}
\usage{
build_mcmc_post(comp_data, inifiles, pcs = NULL, smoothing = 0.1,
  cal_mean = TRUE, lowcol = "mina", hicol = "maxb",
  prior_params = NULL, use_lnorm_ecs = TRUE, verbose = FALSE)
}
\arguments{
\item{comp_data}{Table of esm summary statistics (mean, min, max) by year.}

\item{inifiles}{Named vector of names of Hector input files for the various
experiments to be run. You do not need to provide ini files for historical
experiments unless they are the only experiments present.  Extraneous files
will be ignored, so it is safe to use a canoncial list of filenames.}

\item{pcs}{Principal components structure.  If \code{NULL}, work directly
with output values.}

\item{smoothing}{Smoothing parameter for the mesa function, expressed as a
fraction of the window size.}

\item{cal_mean}{If true calibrate to mean; otherwise calibrate to range}

\item{lowcol}{Column in comparison data to use for the low edge of the mesa
function.  Ignored if cal_mean is \code{TRUE}}

\item{hicol}{Column in comparison data to use for the high edge of the mesa
function.  Ignored if cal_mean is \code{TRUE}}

\item{prior_params}{Named list of alternative values for the numerical
parameters in the prior distributions.  Any parameters not mentioned in the
list will be set to their default values.}

\item{use_lnorm_ecs}{If true, use a log-normal prior for climate
sensitivity.  If false, use a normal prior.  The false setting is intended
primarily for testing the influence of priors on the final result.}

\item{verbose}{If \code{TRUE}, print diagnostic messages.  Otherwise,
diagnostic messages will not be generated.}
}
\description{
The function constructed by this call will be suitable for passing to either
a markov chain sampler or an optimizer.  (Though the optimizer will need to
be configured to find a maximum, rather than a minimum.)
}
\details{
Note that the \code{use_c_cycle} argument only controls whether carbon cycle
parameters are sampled and whether the co2 variable is considered in the
likelihood function.  Whether Hector's carbon cycle runs is controled by the
input file \code{inifile}.  It is the user's responsibility to supply a
compatible input file, as well as comparison data that was run with the
desired carbon cycle settings.
}
\section{Notes}{


For some reason we have made the mesa function's \code{sig} parameter
settable for co2, but fixed at 0.4 for temperature.
}

