% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleESMcalibration.R
\name{make_minimize_function}
\alias{make_minimize_function}
\title{Make the function to minimize}
\usage{
make_minimize_function(hector_cores, esm_data, normalize, param,
  cmip_range = NULL, n = NULL, showMessages = FALSE,
  intermediateOutput = FALSE)
}
\arguments{
\item{hector_cores}{A list of Hector cores that will be used.}

\item{esm_data}{A data frame of ESM data for a single model that contains the following columns, year, model, variable, experiment, weight.}

\item{normalize}{A list of center and the scale values to use to noramlize the Hector and ESM output data.}

\item{param}{A vector of parameter values.}

\item{cmip_range}{Default set to NULL and the make_minim_function will only look at the error between esm_data and Hector data.
But if cmip_range is set to a dataframe containing columns (variable, year, sig, lower, and upper) then the minimize function will also
minimize the value returned by the -log of the mesa function.}

\item{n}{The number of cores to parallelize the Hector runs over, unless sepcified will use the number of cores detected by \code{detectCores}}

\item{showMessages}{Default set to FALSE, will supress Hector error messages.}

\item{intermediateOutput}{Default set to FALSE, but if set to TRUE will return the MSE for each variable / experiment / ensemble memeber instead over the over all MSE.}
}
\value{
A function that will calculate the mean squared error between a Hector run and ESM data.
}
\description{
Create the function that will calculate the mean squared error between Hector and ESM CMIP5 output data. THis
function will be minimized using optim.
}
