% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{metrosamp2coda}
\alias{metrosamp2coda}
\title{Convert a list of metrosamp structures to a \code{\link[coda]{mcmc.list}}
structure.}
\usage{
metrosamp2coda(mslist)
}
\arguments{
\item{mslist}{A list mof metrosamp structures}
}
\description{
Convert a list of metrosamp structures to a \code{\link[coda]{mcmc.list}}
structure.
}
