% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{mktruncnorm}
\alias{mktruncnorm}
\title{Create a truncated normal log density function}
\usage{
mktruncnorm(a, b, mu, sig)
}
\arguments{
\item{a}{Lower bound for the truncated normal distribution}

\item{b}{Upper bound for the truncated normal distribution}

\item{mu}{Mean for the untruncated normal distribution the distribution is
derived from.}

\item{sig}{Standard deviation for the untruncated normal distribution the
distribution is derived from.}
}
\value{
A function that computes the log density, suitably normalized.
}
\description{
Given the mean, standard deviation, and cutoffs for a truncated normal PDF,
produce a function that calculates the normalized log density.
}
