% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{parameterize_core}
\alias{parameterize_core}
\title{Set parameters into a hector core}
\usage{
parameterize_core(params, core)
}
\arguments{
\item{params}{A named vector of Hector parameters.}

\item{core}{A Hector core.}
}
\value{
The input core, with parameters reset and spinup rerun.
}
\description{
Set hector parameters by name.  Parameters not mentioned in the input vector
will be left at their current values.  After the new parameters are set, the
core is reset.
}
