% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_varfrac}
\alias{plot_varfrac}
\title{Plot the fraction of variance captured by the PCs}
\usage{
plot_varfrac(pca, nc = NA)
}
\arguments{
\item{pca}{A pca structure}

\item{nc}{Number of components to plot (default is all)}
}
\value{
ggplot dot plot of the fraction of the variance accounted for by
each PC
}
\description{
Plot the fraction of variance captured by the PCs
}
