% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{project_climate}
\alias{project_climate}
\title{Project a data frame of climate data on to a principal componet basis}
\usage{
project_climate(climate_data, principal_components, row_vector = TRUE)
}
\arguments{
\item{climate_data}{a data frame of climate data.  The data must contain
value, year, experiment, and variable columns.}

\item{principal_components}{a \code{hectorcal} object that contains results
from \code{prcomp} and meta data information.}

\item{row_vector}{If \code{TRUE}, return the data as a row vector (a 1xN
matrix)}
}
\value{
a list of three objects, rotation the PC loadings, scale the values
used to scale the input climate data, and center the values used to center
the climate data.
}
\description{
Project a data frame of climate data on to a principal componet basis
}
