% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{project_climate}
\alias{project_climate}
\title{Project a data frame of climate data on to a principal componet basis}
\usage{
project_climate(PCA_climate, principal_components)
}
\arguments{
\item{PCA_climate}{an object of climate data, must contain value, year, experiment, and variable columns.
It may not contain data for more years, variables, or experiments then in the principal_components argument.}

\item{principal_components}{a \code{hectorcal} object that contains results from \code{prcomp} and meta data information.}
}
\description{
Project a data frame of climate data on to a principal componet basis
}
