% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{reconstruct_climate}
\alias{reconstruct_climate}
\title{Reconstruct climate data from projected climate data and prcinicpal componetns}
\usage{
reconstruct_climate(projected_climate, principal_components)
}
\arguments{
\item{projected_climate}{an object returned by \code{project_climate}}

\item{principal_components}{a \code{hectorcal} object that contains results from \code{prcomp} and meta data information.}
}
\description{
Reconstruct climate data from projected climate data and prcinicpal componetns
}
