% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleESMcalibration.R
\name{singleESM_calibration}
\alias{singleESM_calibration}
\title{Emulate a single ESM by calibrating Hector to ESM output data.}
\usage{
singleESM_calibration(inifiles, hector_names, esm_data, normalize,
  initial_param, maxit = 500, n_parallel = NULL,
  showMessages = FALSE)
}
\arguments{
\item{inifiles}{A vector of the Hector inifile cores to use, there should be a core for each esm CMIP comparison data.}

\item{hector_names}{A vector of the Hector core names, should reflect experiment names in the esm CMIP comparion data.}

\item{esm_data}{A data frame of ESM data for a single model that contains the following columns, year, model, variable, experiment.}

\item{normalize}{A list of center and the scale values to use to noramlize the Hector and ESM output data.}

\item{initial_param}{A named vector of inital paramters to be optimized over.}

\item{maxit}{The max number of itterations for optim, default set to 500.}

\item{n_parallel}{The max number of cores to parallize the runs over,  unless sepcified will use the number of cores detected by \code{detectCores}.}

\item{showMessages}{Default set to FALSE, will supress Hector error messages.}
}
\value{
An object returned by \code{optim}
}
\description{
Emulate a single ESM by calibrating Hector to ESM output data.
}
