% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleESMcalibration.R
\name{singleESM_calibration_diag}
\alias{singleESM_calibration_diag}
\title{Emulate a single ESM by calibrating Hector to ESM output data and return diagnostic materials.}
\usage{
singleESM_calibration_diag(inifiles, hector_names, esm_data, normalize,
  initial_param, core_weights = NULL, maxit = 500, n_parallel = NULL)
}
\arguments{
\item{inifiles}{A vector of the Hector inifile cores to use, there should be a core for each esm CMIP comparison data.}

\item{hector_names}{A vector of the Hector core names, should reflect experiment names in the esm CMIP comparion data.}

\item{esm_data}{A data frame of ESM data for a single model that contains the following columns, year, model, variable, experiment.}

\item{normalize}{A list of center and the scale values to use to noramlize the Hector and ESM output data.}

\item{initial_param}{A named vector of inital paramters to be optimized over.}

\item{core_weights}{An optional vector of experiment weights to get the weighted sum of experiment MSE, the default it set to NULL the no weights are used.}

\item{maxit}{The max number of itterations for optim, default set to 500.}

\item{n_parallel}{The max number of cores to parallize the runs over,  unless sepcified will use the number of cores detected by \code{detectCores}.}
}
\value{
A list containing the following elements, copmarison_plot a plot comparing Hector and ESM output data, residual_plot a plot comparing the
normalized residuals, MSE a data frame of the mean squared error for each experiment and variable optim minimizes the sum of the MSE values, and
optim_rslt is the object returned by \code{optim}.
}
\description{
Calibrate Hector so that is emulates a single ESM. Calcaulte the mean squared error between the calibrated Hector and the
ESM comparison data, plot the Hector output vs the ESM comparison data, and a plot of the residuals.
}
