% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleESMcalibration.R
\name{translate_variable_name}
\alias{translate_variable_name}
\title{Switch from Hector variable names to ESM variable names.}
\usage{
translate_variable_name(input)
}
\arguments{
\item{input}{A dataframe of climate data that contains variable names that have to be transalted.}
}
\value{
A dataframe of Hector output data but with the ESM variable names.
}
\description{
Hector and ESM use different nomenclature for the variables, this function changes the Hector
variable name to the esm variable name for (tas, co2, and heatflux but other variables may be added).
}
